More robust Clip-ID determination
Python's glob may not return the list of files in the order
assumed by the previous implementation. This new impl does
not assume order and instead walks the entire list of parquet
files searching for one of the required ones to determine
clip-id. This issue was found and triggered by the new
world scenario video generation doc. The new example asset
bundling triggered the code path in question which revealed
the bug.

Signed-off-by: ccasanova <ccasanova@nvidia.com>