Batch frames to GPU for face blur
Customer was hitting OOM on 15s video during the face blur step.
The code was putting the entire video onto the GPU which could
exhaust VRAM even on an 80GB H100.

This change now loads only a batch's worth of frames to the GPU
at a time thereby keeping peak GPU memory much, much lower (and
no OOM of course!).

Signed-off-by: ccasanova <ccasanova@nvidia.com>